feat: 신고하기 코드 연결 but 안뜸 #28

+++ app/src/main/java/com/eatssu/android/data/model/request/ReportRequest.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ReportRequest(
     @SerializedName("reviewId")
-    val reviewId: Int,
+    val reviewId: Long,
 
     @SerializedName("reportType")
     val reportType: String,
